Move conversation classes to their own package.
