+++ res/layout/conversation_item_received.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<org.thoughtcrime.securesms.ConversationItem
+<org.thoughtcrime.securesms.conversation.ConversationItem
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -193,4 +193,4 @@
             android:gravity="center_vertical"/>
 
     </RelativeLayout>
-</org.thoughtcrime.securesms.ConversationItem>
+</org.thoughtcrime.securesms.conversation.ConversationItem>
