+++ res/layout/conversation_item_update.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<org.thoughtcrime.securesms.ConversationUpdateItem
+<org.thoughtcrime.securesms.conversation.ConversationUpdateItem
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/conversation_update_item"
@@ -77,4 +77,4 @@
 
     </LinearLayout>
 
-</org.thoughtcrime.securesms.ConversationUpdateItem>
+</org.thoughtcrime.securesms.conversation.ConversationUpdateItem>
