+++ src/org/thoughtcrime/securesms/ShareActivity.java
@@ -32,8 +32,10 @@ import android.support.annotation.Nullable;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.app.ActionBar;
 import android.support.v7.widget.Toolbar;
+
+import org.thoughtcrime.securesms.conversation.ConversationActivity;
 import org.thoughtcrime.securesms.logging.Log;
-import android.view.MenuItem;
+
 import android.view.View;
 import android.widget.ImageView;
